comment-list-item 용도별 폴더 분기함
